feat: Init app file with splash screen
