+++ package.json
@@ -16,6 +16,7 @@
     "expo": "~50.0.13",
     "expo-build-properties": "^0.11.1",
     "expo-dev-client": "^3.3.10",
+    "expo-splash-screen": "^0.26.4",
     "expo-status-bar": "~1.11.1",
     "react": "18.2.0",
     "react-native": "0.73.6"
