+++ src/app.tsx
@@ -1,0 +1,36 @@
+import { StatusBar } from 'expo-status-bar';
+import React from 'react';
+import { StyleSheet, View, Text } from 'react-native';
+
+interface AppProps {
+  hideSplashScreen: () => Promise<boolean>;
+}
+
+/**
+ * This is the root component of our app.
+ * @param {AppProps} props - The props for the `App` component.
+ * @returns {JSX.Element} The rendered `App` component.
+ */
+function App(props: AppProps): JSX.Element {
+  const { hideSplashScreen } = props;
+
+  setTimeout(hideSplashScreen, 500);
+
+  return (
+    <View style={styles.container}>
+      <Text>Open up App.tsx to start working on your app!</Text>
+      <StatusBar style="auto" />
+    </View>
+  );
+}
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    backgroundColor: '#fff',
+    alignItems: 'center',
+    justifyContent: 'center',
+  },
+});
+
+export default App;
